requirements.txt
fastapi
uvicorn[standard]
python-multipart
ebooklib
beautifulsoup4
lxml
aiosqlite
pymupdf
mobi
striprtf
python-docx
rarfile
Pillow
